Move inline trait inlining to its own phase
