+++ compiler/src/dotty/tools/dotc/transform/Inlining.scala
@@ -11,17 +11,12 @@ import dotty.tools.dotc.ast.Trees._
 import dotty.tools.dotc.quoted._
 import dotty.tools.dotc.inlines.Inlines
 import dotty.tools.dotc.ast.TreeMapWithImplicits
-import dotty.tools.dotc.core.DenotTransformers.SymTransformer
 import dotty.tools.dotc.staging.StagingLevel
-import dotty.tools.dotc.core.SymDenotations.SymDenotation
-import dotty.tools.dotc.core.StdNames.str
-import dotty.tools.dotc.core.Types.*
-import dotty.tools.dotc.core.Names.Name
 
 import scala.collection.mutable.ListBuffer
 
 /** Inlines all calls to inline methods that are not in an inline method or a quote */
-class Inlining extends MacroTransform, SymTransformer {
+class Inlining extends MacroTransform {
 
   import tpd._
 
@@ -62,56 +57,6 @@ class Inlining extends MacroTransform, SymTransformer {
       new InliningTreeMap().transform(tree)
   }
 
-  override def transformSym(symd: SymDenotation)(using Context): SymDenotation =
-    if symd.isClass && symd.owner.isInlineTrait && !symd.is(Module) then
-      symd.copySymDenotation(name = newInnerClassName(symd.name), initFlags = (symd.flags &~ Final) | Trait)
-    else
-      symd
-
-  def transformInlineTrait(inlineTrait: TypeDef)(using Context): TypeDef =
-    val tpd.TypeDef(_, tmpl: Template) = inlineTrait: @unchecked
-    val body1 = tmpl.body.flatMap {
-      case innerClass: TypeDef if innerClass.symbol.isClass =>
-        val newTrait = makeTraitFromInnerClass(innerClass)
-        val newType = makeTypeFromInnerClass(inlineTrait.symbol, innerClass, newTrait.symbol)
-        List(newTrait, newType)
-      case member: MemberDef =>
-        List(member)
-      case _ =>
-        // Remove non-memberdefs, as they are normally placed into $init()
-        Nil
-    }
-    val tmpl1 = cpy.Template(tmpl)(body = body1)
-    cpy.TypeDef(inlineTrait)(rhs = tmpl1)
-
-  private def makeTraitFromInnerClass(innerClass: TypeDef)(using Context): TypeDef =
-    val TypeDef(name, tmpl: Template) = innerClass: @unchecked
-    val newInnerParents = tmpl.parents.mapConserve(ConcreteParentStripper.apply)
-    val tmpl1 = cpy.Template(tmpl)(parents = newInnerParents) // TODO .withType(???)
-    val newTrait = cpy.TypeDef(innerClass)(name = newInnerClassName(name), rhs = tmpl1)
-    newTrait.symbol.setFlag(Synthetic)
-    newTrait
-  end makeTraitFromInnerClass
-
-  private def makeTypeFromInnerClass(parentSym: Symbol, innerClass: TypeDef, newTraitSym: Symbol)(using Context): TypeDef =
-    val upperBound = innerClass.symbol.primaryConstructor.info match {
-      case _: MethodType =>
-        newTraitSym.typeRef
-      case poly: PolyType =>
-        HKTypeLambda(poly.paramNames)(tl => poly.paramInfos, tl => newTraitSym.typeRef.appliedTo(tl.paramRefs.head))
-    }
-    val newTypeSym = newSymbol(
-      owner = parentSym,
-      name = newTraitSym.name.asTypeName,
-      flags = innerClass.symbol.flags & (Private | Protected) | Synthetic,
-      info = TypeBounds.upper(upperBound),
-      privateWithin = innerClass.symbol.privateWithin,
-      coord = innerClass.symbol.coord,
-      nestingLevel = innerClass.symbol.nestingLevel,
-    ).asType
-    TypeDef(newTypeSym)
-  end makeTypeFromInnerClass
-
   private class InliningTreeMap extends TreeMapWithImplicits {
 
     /** List of top level classes added by macro annotation in a package object.
@@ -121,13 +66,6 @@ class Inlining extends MacroTransform, SymTransformer {
 
     override def transform(tree: Tree)(using Context): Tree = {
       tree match
-        case tree: TypeDef if tree.symbol.isInlineTrait =>
-          transformInlineTrait(tree)
-        case tree: TypeDef if Inlines.needsInlining(tree) =>
-          val tree1 = super.transform(tree).asInstanceOf[TypeDef]
-          if tree1.tpe.isError then tree1
-          else if tree1.symbol.isInlineTrait then transformInlineTrait(tree1)
-          else Inlines.inlineParentInlineTraits(tree1)
         case tree: MemberDef =>
           if tree.symbol.is(Inline) then tree
           else if tree.symbol.is(Param) then super.transform(tree)
@@ -167,18 +105,6 @@ class Inlining extends MacroTransform, SymTransformer {
         case _ =>
           if tree.isType then tree
           else super.transform(tree)
-    }
-  }
-
-  private def newInnerClassName(name: Name): name.ThisName = name ++ str.INLINE_TRAIT_INNER_CLASS_SUFFIX
-
-  private object ConcreteParentStripper extends TreeAccumulator[Tree] {
-    def apply(tree: Tree)(using Context): Tree = apply(tree, tree)
-
-    override def apply(x: Tree, tree: Tree)(using Context): Tree = tree match {
-      case ident: Ident => ident
-      case tpt: TypeTree => tpt
-      case _ => foldOver(x, tree)
     }
   }
 }
